+++ src/gui_child_manager.cpp
@@ -64,23 +64,23 @@ void
 GUIChildManager::draw(GraphicContext& parent_gc)
 {
   GCZoomState zoom(0, 0, parent_gc.get_width(), parent_gc.get_height());
-  zoom.set_offset(static_cast<float>(m_x), static_cast<float>(m_y));
+  zoom.set_offset(m_x, m_y);
 
   ZoomGraphicContext gc(parent_gc, zoom);
 
-  parent_gc.draw_fill_rect(static_cast<float>(m_x), static_cast<float>(m_y),
-                           static_cast<float>(m_x + m_width), static_cast<float>(m_y + m_height),
+  parent_gc.draw_fill_rect(m_x, m_y,
+                           m_x + m_width, m_y + m_height,
                            Colors::button_bg_passive);
-  parent_gc.draw_rect(static_cast<float>(m_x), static_cast<float>(m_y),
-                      static_cast<float>(m_x + m_width), static_cast<float>(m_y + m_height),
+  parent_gc.draw_rect(m_x, m_y,
+                      m_x + m_width, m_y + m_height,
                       Colors::button_fg_passive);
 
-  parent_gc.draw_fill_rect(static_cast<float>(m_x), static_cast<float>(m_y),
-                           static_cast<float>(m_x + m_width), static_cast<float>(m_y),
+  parent_gc.draw_fill_rect(m_x, m_y,
+                           m_x + m_width, m_y,
                            Colors::button_bg_hover);
 
-  parent_gc.draw_rect(static_cast<float>(m_x), static_cast<float>(m_y),
-                      static_cast<float>(m_x + m_width), static_cast<float>(m_y),
+  parent_gc.draw_rect(m_x, m_y,
+                      m_x + m_width, m_y,
                       Colors::button_fg_passive);
 
   for (auto i = m_components.rbegin (); i != m_components.rend (); ++i)
